requirements_el.txt
absl-py==2.1.0
accelerate==1.6.0
ace_tools==0.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
alembic==1.13.2
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.6.2.post1
aplr==10.8.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
astor==0.8.1
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.5
async-timeout==4.0.3
attrs==24.2.0
autograd==1.7.0
autograd-gamma==0.5.0
autonormalize==2.0.1
babel==2.17.0
backoff==2.2.1
bcrypt==4.2.0
beautifulsoup4==4.13.3
bleach==6.2.0
blinker==1.8.2
build==1.2.2.post1
cachetools==5.5.0
catboost==1.2.7
category-encoders==2.6.3
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
chroma-hnswlib==0.7.6
chromadb==0.5.17
click==8.1.8
cloudpickle==3.1.0
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.9.0
comm==0.2.2
commonmark==0.9.1
contourpy==1.3.0
cycler==0.12.1
dash==2.18.2
dash-core-components==2.0.0
dash-cytoscape==1.0.2
dash-html-components==2.0.0
dash-table==5.0.0
dask==2025.3.0
deap==1.4.2
debugpy==1.8.5
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
dill==0.3.9
distributed==2025.3.0
distro==1.9.0
dnspython==2.7.0
docx2txt==0.8
durationpy==0.9
einops==0.8.1
email_validator==2.2.0
et_xmlfile==2.0.0
exceptiongroup==1.2.2
executing==2.1.0
Faker==36.1.1
fastapi==0.115.4
fastjsonschema==2.20.0
feather-format==0.4.1
featuretools==1.31.0
featurewiz==0.6.1
filelock==3.16.1
Flask==3.0.3
flask-cors==5.0.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
flatbuffers==24.3.25
fonttools==4.53.1
formulaic==1.1.1
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2024.10.0
gast==0.6.0
gevent==24.11.1
google-auth==2.35.0
google-pasta==0.2.0
googleapis-common-protos==1.65.0
graphviz==0.20.3
greenlet==3.1.1
groq==0.11.0
grpcio==1.67.1
h11==0.14.0
h2o==3.46.0.7
h5py==3.12.1
holidays==0.69
httpcore==1.0.6
httptools==0.6.4
httpx==0.27.2
huggingface-hub==0.26.2
humanfriendly==10.0
hyper-connections==0.1.8
idna==3.10
imbalanced-learn==0.13.0
imblearn==0.0
importlib_metadata==8.4.0
importlib_resources==6.4.5
interface-meta==1.3.0
interpret==0.6.9
interpret-core==0.6.9
ipykernel==6.29.5
ipython==8.27.0
ipywidgets==8.1.5
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
json5==0.12.0
jsonify==0.5
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
keras==3.8.0
kiwisolver==1.4.7
kubernetes==31.0.0
langchain==0.3.7
langchain-chroma==0.1.4
langchain-core==0.3.15
langchain-groq==0.2.1
langchain-huggingface==0.1.2
langchain-text-splitters==0.3.2
langsmith==0.1.139
lazytransform==1.18
libclang==18.1.1
lifelines==0.30.0
lightgbm==4.5.0
lightning==2.5.1
lightning-utilities==0.11.9
llvmlite==0.43.0
locket==1.0.0
logitboost==0.7
loky==3.5.1
Mako==1.3.5
mambular==1.4.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.3
ml-dtypes==0.4.1
mlrose_hiive==2.2.4
mlxtend==0.23.4
mmh3==5.0.1
monotonic==1.6
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
multiprocess==0.70.17
mysql-connector-python==9.0.0
namex==0.0.8
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
ngboost==0.5.1
node==1.2.2
notebook==7.3.3
notebook_shim==0.2.4
numba==0.60.0
numexpr==2.10.2
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
oauthlib==3.2.2
odict==1.9.0
omegaconf==2.3.0
onnxruntime==1.20.0
openpyxl==3.1.5
opentelemetry-api==1.27.0
opentelemetry-exporter-otlp-proto-common==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
opentelemetry-instrumentation==0.48b0
opentelemetry-instrumentation-asgi==0.48b0
opentelemetry-instrumentation-fastapi==0.48b0
opentelemetry-proto==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
opentelemetry-util-http==0.48b0
opt_einsum==3.4.0
optree==0.14.0
optuna==4.1.0
orjson==3.10.11
overrides==7.7.0
packaging==24.1
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
partd==1.4.2
patsy==0.5.6
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.2.2
plotly==5.24.0
plumber==1.7
posthog==3.7.0
prometheus_client==0.21.1
prompt_toolkit==3.0.47
propcache==0.2.0
properscoring==0.1
protobuf==4.25.5
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyaml==25.1.0
pyants==0.3.0
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
PyMySQL==1.1.1
pyparsing==3.1.4
PyPDF2==3.0.1
PyPika==0.48.9
pyproject_hooks==1.2.0
pyswarm==0.6
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==3.3.0
pytorch-lightning==2.1.4
pytorch-tabnet==4.1.0
pytorch-tabular==1.1.1
pytz==2024.1
PyYAML==6.0.2
pyzmq==26.2.0
qhoptim==1.1.0
ray==2.40.0
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.20.0
rsa==4.9
safetensors==0.4.5
SALib==1.5.1
scikit-learn==1.5.2
scikit-optimize==0.10.2
scipy==1.15.2
seaborn==0.13.2
Send2Trash==1.8.3
sentence-transformers==3.2.1
shap==0.46.0
shellingham==1.5.4
six==1.16.0
sklearn-compat==0.1.3
slicer==0.0.8
snapboost==0.1.2
snapml==1.16.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.34
stack-data==0.6.3
starlette==0.41.2
statsmodels==0.14.2
sympy==1.13.1
tab-transformer-pytorch==0.4.1
tabpfn==2.0.8
tabulate==0.9.0
tblib==3.1.0
tenacity==9.0.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.5.0
terminado==0.18.1
tf_keras==2.18.0
threadpoolctl==3.5.0
tinycss2==1.4.0
tokenizers==0.20.2
tomli==2.2.1
toolz==1.0.0
torch==2.5.1
torchmetrics==1.5.2
tornado==6.4.1
tqdm==4.66.6
traitlets==5.14.3
transformers==4.46.1
triton==3.1.0
typer==0.9.0
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2024.1
uri-template==1.3.0
urllib3==2.2.3
uvicorn==0.32.0
uvloop==0.21.0
watchfiles==0.24.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==13.1
Werkzeug==3.0.4
widgetsnbextension==4.0.13
woodwork==0.31.0
wrapt==1.16.0
WTForms==3.1.2
xgboost==1.7.6
xlrd==2.0.1
yacs==0.1.8
yarl==1.17.1
zict==3.0.0
zipp==3.20.2
zope.component==6.0
zope.deferredimport==5.0
zope.deprecation==5.1
zope.event==5.0
zope.hookable==7.0
zope.interface==7.2
zope.lifecycleevent==5.1
zope.proxy==6.1
